Feature/websocket comms (#163)
This PR provides the scaffolding for single connection communication between the UI and the backend (sitting on the raspberry pi)

Leaving the logs for now but we should remove some chatty ones before final release (which is gonna take a while)

Inception
The FE and BE commincation will be in a biderectional Command => Event driven approach.
The FE implementation splits the websocket (on inteface level) into a source (for reading the events) and a sink (for sending the commands) .
A sink is essentially a consumer and a source a producer. In this case the interface is exposing the Source for consumers (component views) to process events and the sink for sending commands (i.e. with a service call interfacing a component) (thus replacing http calls )