Reboot and shutdown by issuing syscalls instead of calling shell commands (#174)
Currently the poweroff function inherits the priviledges of the gui since we are running this from the e.g. gnome-terminal and poweroff and reboot work without needing root priviledges as they inherit the gui ones (I think). However when it runs as a service this will not apply.

Go has also a linux syscall API that we can use instead of issuing shell commands which will dramatically decrease response times (although not a big deal for power management, having this paradigm in would be beneficial for later)